lambda/just_in_time/requirements.txt
jsonschema
boto3
auth0-python